+++ src/main/java/com/bombombom/devs/user/repository/UserRepository.java
@@ -1,9 +1,12 @@
-package com.bombombom.devs.user;
+package com.bombombom.devs.user.repository;
 
 import com.bombombom.devs.user.models.User;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findUserByUsername(String username);
 
     void deleteByUsername(String username);
